% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynprog.R
\name{DYNPROG}
\alias{DYNPROG}
\title{Optimal segmentation with dynamic programming}
\usage{
DYNPROG(dataVec, maxSegs)
}
\arguments{
\item{dataVec}{A vector of data to be segmented}

\item{maxSegs}{The maximum number of segments to create}
}
\value{
A cost matrix with the entry at (i, j) being the optimal cost
for creating i segments up to data point j
}
\description{
Optimal segmentation with dynamic programming
}
\examples{
sampleVec <- 1:30
result <- DYNPROG(sampleVec, 5)
}
