% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Title}
\usage{
PCA(data.matrix)
}
\arguments{
\item{data.matrix}{The feature matrix on which to perform PCA}
}
\value{
A rotation matrix, a matrix of loadings, and a center matrix
}
\description{
Title
}
\examples{
PCA(iris[,2:3])
}
