% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynprog.R
\name{DYNPROG}
\alias{DYNPROG}
\title{Title}
\usage{
DYNPROG(dataVec, maxSegs)
}
\arguments{
\item{dataVec}{A vector of data to be segmented}

\item{maxSegs}{The maximum number of segments to create}
}
\value{
A cost matrix with the entry at (i, j) being the optimal cost
for creating i segments up to data point j
}
\description{
Title
}
\examples{
if(!file.exists("data-for-LOPART-signals.csv.gz"))
{
 download.file(
   "https://raw.githubusercontent.com/tdhock/LOPART-paper/master/data-for-LOPART-labels.csv.gz",
   "data-for-LOPART-signals.csv.gz")
}
signals.dt <- data.table::fread("data-for-LOPART-signals.csv.gz")
singleSeq.dt <- signals.dt[sequenceID=="20167.22",]
DYNPROG(singleSeq.dt$logratio, 5)
}
