% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynprog.R
\name{getOptimalCost}
\alias{getOptimalCost}
\title{Title}
\usage{
getOptimalCost(qVec, sVec, start, end)
}
\arguments{
\item{qVec}{Sum of squares up to the end of the data}

\item{sVec}{Sum of the data}

\item{start}{Beginning index}

\item{end}{Ending index}
}
\value{
A vector of optimal costs
}
\description{
Title
}
\examples{
testVec <- 1:30
qVec <- cumsum(testVec^2)
sVec <- cumsum(testVec)
cost <- getOptimalCost(qVec, sVec, 1, 4)
}
